video: host: tsec: make message handling efficient
For message handling we ack the interrupt from tsec in topHalf and
consume the messages in bottomHalf. In this fix we also mask the
interrupt in topHalf so that any subsequent interrupts generated by tsec
doesn't enter the topHalf unless bottomHalf has finished consuming the
current round of messages. Once bottomHalf is done we unmask the
interrupt and if it is pending then topHalf will be invoked again.

Bug 3381288

Change-Id: Ib06d4805c2f421d2deb56c00bae5482c08ae2c94
Signed-off-by: Nikesh Oswal <noswal@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2636053
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>