video: host: tsec: boot tsec on kernel boot
Usecase for TSEC is that it will be always on and will be rebooted only
during system level suspend/resume. There is no client of TSEC on CCPLEX
but TSEC firmware directly talks with DCE firmware. DCE firmware is
booted by MB1 and in this patch we add support in TSEC driver to boot
TSEC when kernel boots.

Bug 3381288

Change-Id: I22baf5661602cc08f2be1dc3023873d9e30425d7
Signed-off-by: Nikesh Oswal <noswal@nvidia.com>
Signed-off-by: Sahil Mukund Patki <spatki@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2719165
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>