drivers: pva: remove add/remove nvhost client
nvhost_module_remove_client takes ~200ms when the PVA
is power gated. This occurs due to BPMP first powering on
the PVA to update the clocks and then powering it off.

Since PVA does not support DVFS, there is no need to
add the pva as a client to nvhost.

This results in ~4ms for open/release of PVA device.

Bug 3430361

Change-Id: Ib37ae1aea7f451d9902af6d94b7e61da574d3e89
Signed-off-by: omar nemri <onemri@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2715702
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>