drivers: pva: fix first authenticate after power up
After powerup, the first authetication of an invalid app
suceeds although it is suppost to fail.

This happened because the system allow list was not
enabled until it is parsed and that happens on finalize
power on wich happens after task submit.  Authentication
happens during register elf.

Bug 3695575

Change-Id: Iaeb9d3a9a640f2635ad65dacf00872b22f5a1739
Signed-off-by: omar nemri <onemri@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2735240
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
GVS: Gerrit_Virtual_Submit