nvhost: pva: add support for pva clock capping
For K5.15 the nvhost_acm driver is not used and thus the central
mechanism for clock capping for the nvhost cliets like PVA is not
available.

The clock capping functionality should be part of the individual
nvhost drivers from K5.15 onwards and thus adding the related
support in nvhost-pva drivers.

Bug 4029003

Change-Id: I005008c0072a527f427ccb4eadbc5e83338a38eb
Signed-off-by: Ninad Malwade <nmalwade@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2899032
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-cert <svc-mobile-cert@nvidia.com>
Reviewed-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>